% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_recodes_39.R
\name{recode.cause.39}
\alias{recode.cause.39}
\title{ICD-10 Index Of 39 Selected Causes of Death}
\usage{
recode.cause.39(icd10)
}
\arguments{
\item{icd10}{character ICD-10 code}
}
\value{
Recode value for 39 Selected Causes of Death
}
\description{
ICD-10 Index Of 39 Selected Causes of Death
}
\examples{
recode.cause.39("C45")

}

